refactor: useTable, usePagination
