+++ app/[lang]/code/_components/server-only/GroupCodeTable.tsx
@@ -12,8 +12,8 @@ import {
 import { Pagination } from '@nextui-org/pagination'
 import { Spinner } from '@nextui-org/spinner'
 import { fetcherJson } from '@/app/_utils/fetch'
-import { pagination } from '@/app/_components/server-only/primitives'
-import useCmTable, { getCmPagination } from './table'
+import useCmTable from '../../../../_components/server-only/table/use'
+import getCmPaginationProps from '@/app/_components/server-only/pagination/use'
 
 export default function GroupCodeTable(props: {
   page: number
@@ -22,6 +22,7 @@ export default function GroupCodeTable(props: {
   classNames?: TableProps['classNames']
 }) {
   const { page, setPage, handleSelect } = props
+
   const { tableProps } = useCmTable({
     tableProps: {
       'aria-label': 'Group Code Table',
@@ -37,23 +38,18 @@ export default function GroupCodeTable(props: {
   )
   if (!data) return <div>Loading...</div>
   const { totalPage, data: bodyData, currentPage } = data.body
+  const { paginationProps } = getCmPaginationProps({
+    page: currentPage ?? 0,
+    total: totalPage ?? 0,
+    onChange: (page) => setPage && setPage(page),
+  })
   const loadingState = isLoading || bodyData?.length === 0 ? 'loading' : 'idle'
 
   return (
     <Table
       {...tableProps}
       key={`group-code-table-${page}`}
-      bottomContent={
-        totalPage > 0 ? (
-          <Pagination
-            {...getCmPagination({
-              page: currentPage,
-              total: totalPage,
-              onChange: (page) => setPage && setPage(page),
-            }).paginationProps}
-          />
-        ) : null
-      }
+      bottomContent={totalPage > 0 ? <Pagination {...paginationProps} /> : null}
     >
       <TableHeader key={'group-code-table-header'}>
         <TableColumn key="codeGroup">그룹코드</TableColumn>
